pyproject.toml
[project]
name = "dns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "octodns>=1.13.0",
    "octodns-cloudflare>=1.0.0",
    "pyyaml>=6.0.0",
]
